move cookies into /src/actions
