+++ src/app/actions/actions.ts
@@ -5,16 +5,6 @@ import { savefeedbackToGoogleSheet } from "./email";
 import { postCodeRegex } from "../lib/application";
 import { cookies } from "next/headers";
 
-export async function createCookies(value: any) {
-  const cookieStore = cookies();
-  cookieStore.set("isShowCookie", "false"),
-    cookieStore.set("isConsentCookie", value.toString(), {
-      path: "/",
-      maxAge: 31536000,
-      sameSite: "strict",
-    });
-}
-
 export async function saveComments(data: any) {
   try {
     const { applicationNumber, feeling, comment, postcode } = data;
